Day 09: Parts one and two solutions
This was quite simple; I just implemented the algorithm as described by
the puzzle statement. Usually, in AoC, a naive implementation of the
described algorithm is too inefficient for part two, requiring one to
come up with a smarter algorithm.
For this puzzle, though, the part two solution is as basically the same
as the part one solution, at least in terms of computational complexity.
So there are probably smart optimizations one can make here, but it runs
fine as-is.
This puzzle was a neat illustration of what a derivative of a function
is.